requirements.txt
fastapi==0.109.2
requests>=2.31.0
uvicorn>=0.23.1
pandas==2.2.0
numpy>=1.24.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
pytest==8.3.4
python-dotenv==1.0.1
openai==1.59.7
httpx==0.24.1
python-multipart==0.0.6
grpcio==1.68.0
jina==3.28.0